refactor: notification - reader, writer
