+++ src/module/notification/notification.module.ts
@@ -1,7 +1,8 @@
 import { Module } from '@nestjs/common';
 import { NotificationController } from './notification.controller';
 import { NotificationService } from './notification.service';
-import { NotificationRepository } from './repository/notification.repository';
+import { NotificationReader } from './repository/notification.reader';
+import { NotificationWriter } from './repository/notification.writer';
 import { NotificationListener } from './notification.listener';
 import { WsModule } from '../websocket/ws.module';
 
@@ -10,9 +11,9 @@ import { WsModule } from '../websocket/ws.module';
   controllers: [NotificationController],
   providers: [
     NotificationService,
-    NotificationRepository,
+    NotificationReader,
+    NotificationWriter,
     NotificationListener,
   ],
-  exports: [NotificationRepository],
 })
 export class NotificationModule {}
